requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Mail
python-dotenv
PyMySQL
Werkzeug
cryptography
itsdangerous